add controllers, services, ui templates
