+++ requirements.txt
@@ -8,5 +8,22 @@ python_dotenv
 openai-whisper
 ffmpeg-python
 
-torch
-python-dotenv
+torch~=2.0.1
+python-dotenv~=1.0.0
+
+services~=0.1.1
+pip~=22.3.1
+wheel~=0.40.0
+MarkupSafe~=2.1.2
+Werkzeug~=2.3.6
+click~=8.1.6
+blinker~=1.6.2
+Jinja2~=3.1.2
+itsdangerous~=2.1.2
+pluggy~=1.2.0
+exceptiongroup~=1.1.2
+iniconfig~=2.0.0
+numpy~=1.24.3
+setuptools~=65.5.1
+future~=0.18.3
+zipp~=3.16.2
