+++ src/templates/index.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Meeting Summarizer</title>
+</head>
+<body>
+
+</body>
+</html>
